requirements.txt
autogen==0.9.6
google-generativeai>=0.3.0
streamlit
python-dotenv
SpeechRecognition
PyAudio
pyttsx3
rich